--- environments/mcp_env/mcp_env/__init__.py
@@ -1,3 +1,0 @@
-from .mcp_env import load_environment
-
-__all__ = ["load_environment"]
