--- environments/mcp_env/mcp_env/models.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-from typing import Dict, List
-
-
-@dataclass
-class MCPServerConfig:
-    name: str
-    command: str
-    args: List[str] | None = None
-    env: Dict[str, str] | None = None
-    description: str = ""
