+++ pyproject.toml
@@ -30,17 +30,18 @@ classifiers = [
 dependencies = [
     "datasets>=3.0.0",
     "jinja2>=3.1.6",
+    "math-verify>=0.8.0",
+    "nest-asyncio>=1.6.0", # for jupyter notebooks
     "openai>=1.108.1",
     "openai-agents>=0.0.7",
-    "nest-asyncio>=1.6.0", # for jupyter notebooks
+    "prime-sandboxes>=0.2.5",
+    "pydantic>=2.11.9",
     "requests",
     "rich",
     "tenacity>=8.5.0",
     "textual",
-    "pydantic>=2.11.9",
     "tomli; python_version < '3.11'",
     "typing_extensions; python_version < '3.12'",
-    "prime-sandboxes>=0.2.5",
     "wget>=3.2",
 ]
 
@@ -58,7 +59,6 @@ dev = [
     "reasoning-gym",
     "textarena",
     "nltk",
-    "math-verify>=0.8.0",
 ]
 
 [project.optional-dependencies]
@@ -76,11 +76,10 @@ rl = [
 ]
 envs = [
     "math-verify>=0.8.0",
-    "duckduckgo-search",
-    "brave-search",
     "reasoning-gym",
     "nltk",
     "textarena",
+    "mcp>=1.14.1",
 ]
 
 [tool.uv.extra-build-dependencies]
